added image deletion from storage

+++ app/src/main/java/com/xlsoft/meldcxtask/ui/features/history/HistoryFragment.kt
@@ -1,6 +1,7 @@
 package com.xlsoft.meldcxtask.ui.features.history
 
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
@@ -10,6 +11,7 @@ import com.xlsoft.meldcxtask.core.ui.BaseFragment
 import com.xlsoft.meldcxtask.databinding.FragmentHistoryBinding
 import com.xlsoft.meldcxtask.ui.features.shared.SharedViewModel
 import dagger.hilt.android.AndroidEntryPoint
+import java.io.File
 
 @AndroidEntryPoint
 class HistoryFragment : BaseFragment<FragmentHistoryBinding>() {
@@ -30,7 +32,14 @@ class HistoryFragment : BaseFragment<FragmentHistoryBinding>() {
             adapter = HistoryListAdapter(
                 historyList = it,
                 deleteCallBack = {searchHistory->
-                    _viewModel.deleteHistory(searchHistory)
+                    val file = File(searchHistory.imagePath)
+                    val deleted: Boolean = file.delete()
+                    if(deleted){
+                        Log.d("HistoryFragment", "Deleted image file from storage")
+                        _viewModel.deleteHistory(searchHistory)
+                    }else{
+                        Log.d("HistoryFragment", "Failed to delete image")
+                    }
                 },
                 itemClickCallBack = {searchHistory->
                     _sharedViewModel.setSelectedHistory(searchHistory)
